tests/wpt/tests/shared-storage/select-url-permissions-policy-self.tentative.https.sub.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/permissions-policy/resources/permissions-policy.js></script>
  <script src="/shared-storage/resources/util.js"></script>
  <script>
    'use strict';
    const same_origin_src = '/shared-storage/resources/select-url-permissions-policy-helper.html';
    const cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
      same_origin_src;
    const header = 'permissions policy header shared-storage-select-url=(self)';

    promise_test(async t => {
      const allowed = await IsSharedStorageSelectUrlAllowedByPermissionsPolicy();
      assert_true(allowed);
    }, header + ' allows sharedStorage.selectURL() in the current page.');

    async_test(t => {
      test_feature_availability('shared-storage-select-url', t, same_origin_src,
          expect_feature_available_default);
    }, header + ' allows sharedStorage.selectURL() in same-origin iframes.');

    async_test(t => {
      test_feature_availability('shared-storage-select-url', t, cross_origin_src,
          expect_feature_unavailable_default);
    }, header + ' disallows sharedStorage.selectURL() in cross-origin iframes.');
  </script>
</body>
